graphapackage.sh: Handle the case of single output files.
Sometimes, only one PDF file would be created, because a package
is only present in one distribution. In such a case, the call to
pdfunite would fail, because it expects at least two sourece
files to unite.

We now collect the names of successfully created PDFs in an array.
At the end, we check the number of items in the array. If there is
only one item in there, we rename that file to be the final output
PDF. For the case of more than one file, the wildcard call to
pdfunite is retained.

Signed-off-by: Thorsten Fischer <thorsten@froschi.org>